Moved the actor counter into the dimension class
