+++ amulet/level/formats/leveldb_world/format.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import os
 import struct
-import threading
 import warnings
 from typing import Tuple, Dict, Union, Optional, List, BinaryIO, Iterable, Any
 from io import BytesIO
@@ -91,48 +90,6 @@ class BedrockLevelDAT(nbt.NBTFile):
             filename_or_buffer.write(buffer.getvalue())
 
 
-class ActorCounter:
-    _lock: threading.RLock
-    _session: int
-    _count: int
-
-    def __init__(self):
-        self._lock = threading.RLock()
-        self._session = -1
-        self._count = 0
-
-    @classmethod
-    def from_level(cls, level: LevelDBFormat):
-        session = level.root_tag.get("worldStartCount", nbt.TAG_Long(0xFFFFFFFF)).value
-        # for some reason this is a signed int stored in a signed long. Manually apply the sign correctly
-        session -= (session & 0x80000000) << 1
-
-        # create the counter object and set the session
-        counter = ActorCounter()
-        counter._session = session
-
-        # increment and write back so there are no conflicts
-        session -= 1
-        if session < 0:
-            session += 0x100000000
-        level.root_tag.value["worldStartCount"] = nbt.TAG_Long(session)
-        level.root_tag.save()
-
-        return counter
-
-    def next(self) -> Tuple[int, int]:
-        """
-        Get the next unique session id and actor counter.
-        Session id is usually negative
-
-        :return: Tuple[session id, actor id]
-        """
-        with self._lock:
-            count = self._count
-            self._count += 1
-        return self._session, count
-
-
 class LevelDBFormat(WorldFormatWrapper):
     """
     This FormatWrapper class exists to interface with the Bedrock world format.
@@ -145,8 +102,6 @@ class LevelDBFormat(WorldFormatWrapper):
     _db: Optional[LevelDB]
     # A class to manage dimension data. This is private
     _dimension_manager: Optional[LevelDBDimensionManager]
-    # A class to keep track of unique actor ids
-    _actor_counter: Optional[ActorCounter]
 
     def __init__(self, path: str):
         """
@@ -163,7 +118,6 @@ class LevelDBFormat(WorldFormatWrapper):
         )
         self._db = None
         self._dimension_manager = None
-        self._actor_counter = None
         self._shallow_load()
 
     def _shallow_load(self):
@@ -274,11 +228,6 @@ class LevelDBFormat(WorldFormatWrapper):
         )
         return self._dimension_manager
 
-    @property
-    def actor_counter(self) -> ActorCounter:
-        self._verify_has_lock()
-        return self._actor_counter
-
     def _get_interface(
         self, raw_chunk_data: Optional[Any] = None
     ) -> BaseLevelDBInterface:
@@ -333,9 +282,8 @@ class LevelDBFormat(WorldFormatWrapper):
         except:
             pass
         try:
-            self._actor_counter = ActorCounter.from_level(self)
             self._db = LevelDB(os.path.join(self.path, "db"))
-            self._dimension_manager = LevelDBDimensionManager(self._db)
+            self._dimension_manager = LevelDBDimensionManager(self)
             self._is_open = True
             self._has_lock = True
             experiments = self.root_tag.get("experiments", {})
